+++ src/_types/IEXRefDataSymbol.d.ts
@@ -1,0 +1,12 @@
+export interface IEXRefDataSymbol {
+	symbol:string
+	name:string
+	date:string
+	type:string
+	iexId:string
+	region:string
+	currency:string
+	isEnabled:boolean
+	figi:string
+	cik:string
+}