+++ src/_types/index.d.ts
@@ -1,3 +1,5 @@
 export * from './fx_rate_params_T'
 export * from './iex_fetch_opts_I'
+export * from './IEXRefDataExchangeSymbol'
+export * from './IEXRefDataSymbol'
 export * from './ticker_o_T'