+++ src/get_ref_data_symbols/index.d.ts
@@ -1,6 +1,5 @@
-import type { FXSymbols } from 'iex-cloud'
-import type { iex_fetch_opts_I } from '../_types'
+import type { iex_fetch_opts_I, IEXRefDataSymbol } from '../_types'
 export declare function get_ref_data_symbols(
 	opts?:iex_fetch_opts_I
-):Promise<[FXSymbols, Response]>;
+):Promise<[IEXRefDataSymbol[], Response]>;
 export { get_ref_data_symbols as get__ref_data_symbols }