+++ src/ref_data_symbols_/index.d.ts
@@ -1,3 +1,2 @@
-import type { FXSymbols } from 'iex-cloud'
-import type { iex_fetch_opts_I } from '../_types'
-export declare function ref_data_symbols_(opts?:iex_fetch_opts_I):Promise<FXSymbols|null>;
+import type { iex_fetch_opts_I, IEXRefDataSymbol } from '../_types'
+export declare function ref_data_symbols_(opts?:iex_fetch_opts_I):Promise<IEXRefDataSymbol[]|null>;